HV: move the ve820 GPU OpRegion address
Move the GPU OpRegion to the tail area of 2G, consecutive to the
acpi data/nvs area.

the benefit of this is that we can treat this area just like acpi
data/nvs when building ve820 table, the code can be rather simply.

...
|<---GPU_OpRegion--->|
|<---ACPI DATA--->|
|<---ACPI NVS--->|
---2G---

This is the prerequisite to next patch.

Signed-off-by: Zhou, Wu <wu.zhou@intel.com>