config-tools: move the pre RTVM's SSRAM base GPA
In order to reunite hpa1_low_part1/2, the SSRAM area has to be
moved to the area next to ACPI and GPU OPREGION.

Thus the guest's RTCT tab, which was origionally duplicated
from host's RTCT tab, has to be modified to reflect this change.

This patch simply uses the board_inspector's RTCT model to parse
the host's RTCT tab, then modifies the SSRAM enties, and saves
it to the dest RTCT bin file.

Signed-off-by: Zhou, Wu <wu.zhou@intel.com>