hv: block guest eist/hwp cpuids and MSRs
This patch is to disable guests from controlling CPU frequency.
Now ACRN has owned CPU frequency control, so guests should no
longer have it. This is done by blocking the eist/hwp cpuids
and MSRs, and as well as the hypercall on p-state info(which the DM
would use to generate guest ACPI p-state related tables).

Alternatively, ACPI p-state can be passed through to VMs if they do not
share pCPUs(or interfere other VM's CPU frequency). Offline tool will
generate a flag to indicate if the ACPI p-state is OK to pass through.
When the flag is set, the eist cpuid/MSRs and hypercalls are not blocked.

Signed-off-by: Wu Zhou <wu.zhou@intel.com>