HV: passthrough pstate to VM if it owns pCPU
Now we are enabling guest to have hybrid cores. Linux kernel can config
CONFIG_SCHED_MC_PRIO to schedule the cores by priority, so that hybrid
cores can get good performance.

Thus, linux kernel need to know about the CPU cores' priority.
In the current kernel it is registered by intel_pstate driver.

Intel_pstate driver relies on HWP and ACPI _CPC to do this job. So we
passed through those to guest. It is not safe to let guests that share
pCPUs to have HWP/pstate, so this is only done when guest is owning
the pCPUs exclucively.

Signed-off-by: Wu Zhou <wu.zhou@intel.com>