hv: bugfix: prevent reschedule when emulating lock
When emulating 0xf0 prefix instruction, the design is to stop other vcpu
from running, then skip the 0xf0 prefix and insert a mtf vmexit on next
instruction. The problem is that before the target vcpu re-enters
the vm and trigger mtf vmexit, it has a chance to schedule out.

This would not only cause other vcpus stall for unnecessary long time,
and more deadly, the vm_lock which is acquired by the split lock
emulation code, would be attempted in other vm’s context(such as
hypercall), causing dead lock. This situation is seen on WAAG sharing
pcpu scenario, causing both WAAG and SOS kernel hang.

The fix is to prevent reschedule when emulating lock. The lock emulation
cycle should be made atomic and completed ASAP.

Signed-off-by: Wu Zhou <wu.zhou@intel.com>