misc: add a camera hal convert
The current Camera HAL interface is implemented in C++ and includes
some classes based on C++ implementation, such as Parameter.

This results in the inability of the C-based acrn-dm to directly access
it.

This patch implements a conversion from C++ to C language, exports it as
a C language symbol, and processes structures such as camera_info and
Parameter that cannot be handled in C language.

Tracked-On: AUTOMOTIVE-213
Signed-off-by: Chenli Wei <chenli.wei@intel.com>