dm: vm_event: add support for RTC change event
When a guest OS performs an RTC change action, we wish this event be
captured by developers, and then they can decide what to do with it.
(e.g., whether to change physical RTC)

There are some facts that makes RTC change event a bit complicated:
- There are 7 RTC date/time regs (year, month…). They can only be
  updated one by one.
- RTC time is not reliable before all date/time regs updated.
- Guests can update RTC date/time regs in any order.
- Guests may update RTC date/time regs during either RTC halted or not
  halted.

But another fact is that when we set date/time in guest OS (e.g., for
Linux and Windows), it usually updates all the 7 date/time regs in one
batch. So this patch developed a method to process and emit RTC change
event:
1. Each RTC date/time reg write triggers an RTC change event containing
   which date/time reg has been changed.
2. The event handler collects those date/time reg writes.
3. If the handler finds out all of the 7 regs has been written in a
   short time window (e.g., 1 second), the guest OS can be considered to
   have just finished an RTC change action. Then DM will try to emit an RTC
   change event(to Libvirt), containint the final time the vRTC has just
   changed to.
4. Events are dropped if the guest doesn't finish writing all 7
   date/time regs in the time window.

Signed-off-by: Wu Zhou <wu.zhou@intel.com>