dm: treate unexpect event from alsa-lib
Streams use `dshare` plugin trigger an xrun when stop PCM if other
shared streams are running.

If the stream has stopped, it will fail to resume. And if the send the event to
FE it will also fail to resume and cause status error. So do not recover a stop stream.

For xfer thread wake up unexpected, one more usleep may cause more release
laytency, add a sem to wake the thread immediatly.

Tracked-On: AUTOMOTIVE-933
Signed-off-by: Yuanyuan Zhao <yuanyuan.zhao@intel.com>