misc: implements camera manager and virtual camera module
This patch implements the most important parts of the camera daemon,
including a camera manager main thread and a virtual camera object
created for each client connected to it, whether that client is an
Service OS normal App or an ACRN-DM, or even a remote object.

The main thread is responsible for connecting and creating virtual
camera objects.

A virtual camera object can receive and forward data from multiple
cameras to the client side.

Tracked-On: AUTOMOTIVE-1330
Signed-off-by: Chenli Wei <chenli.wei@intel.com>